+++ wxc_sdk/workspace_settings/__init__.py
@@ -23,6 +23,7 @@ from ..person_settings.permissions_in import IncomingPermissionsApi
 from ..person_settings.permissions_out import OutgoingPermissionsApi
 from ..person_settings.privacy import PrivacyApi
 from ..person_settings.push_to_talk import PushToTalkApi
+from ..person_settings.voicemail import VoicemailApi
 from ..rest import RestSession
 
 __all__ = ['WorkspaceSettingsApi']
@@ -54,6 +55,7 @@ class WorkspaceSettingsApi(ApiChild, base='workspaces'):
     permissions_out: OutgoingPermissionsApi
     privacy: PrivacyApi
     push_to_talk: PushToTalkApi
+    voicemail: VoicemailApi
 
     def __init__(self, session: RestSession):
         super().__init__(session=session)
@@ -74,3 +76,4 @@ class WorkspaceSettingsApi(ApiChild, base='workspaces'):
         self.permissions_out = OutgoingPermissionsApi(session=session, selector=ApiSelector.workspace)
         self.privacy = PrivacyApi(session=session, selector=ApiSelector.workspace)
         self.push_to_talk = PushToTalkApi(session=session, selector=ApiSelector.workspace)
+        self.voicemail = VoicemailApi(session=session, selector=ApiSelector.workspace)